将HTTPS从FullPipe中独立出来
